requirements.txt
# === CORE ML/AI og TensorFlow GPU ===
absl-py==2.3.1
astunparse==1.6.3
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0
packaging>=20.0,<25.0
protobuf==3.19.6
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.14.1
Werkzeug==3.1.3
wrapt==1.17.2

# === PyTorch & TensorBoard (kræves af din kode) ===
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# === Data Science, Feature Engineering & ML Utils ===
pandas==2.2.2
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
ta==0.11.0
pandas_ta==0.3.14b0
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.5          # <--- Fjern hvis du ikke bruger catboost!
shap==0.44.1             # <--- SHAP for ML explainability
optuna==3.6.1

# === Backtesting, Tabeller, Utility, Logging ===
tabulate==0.9.0
tqdm==4.67.1
schedule==1.2.2
psutil==7.0.0
joblib==1.4.2
gputil==1.4.0             # GPU monitorering

# === Market Data & Trading-API ===
python-binance==1.0.19    # API til Binance data og trading
ccxt==4.1.98              # TILFØJET: CCXT til alle test og fetch_data

# === Telegram, API, CI, Kommunikation ===
python-telegram-bot==20.7
requests==2.32.4
requests-oauthlib==2.0.0
python-dotenv==1.0.1

# === Diverse Utils, I/O, Parsing, Logging ===
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
urllib3==2.5.0

# === Test, udvikling, Jupyter, notebooks, lint ===
pytest==8.4.0
pytest-timeout==2.3.1          # <--- NY! Timeout for pytest
flake8==7.0.0                  # <--- NY! Linting
black==24.4.2                  # <--- NY! Autoformat
mypy==1.10.0                   # <--- NY! Typecheck
coverage==7.5.3                # <--- NY! Test coverage
jupyter==1.0.0
notebook==7.2.0

# === MLflow & DevTools (BRUG KUN DISSE VERSIONER!) ===
mlflow==2.3.0                  # <-- Yngre version pga protobuf
plotly==5.22.0
streamlit==1.16.0              # <-- Yngre version pga protobuf
wandb==0.17.0

# === Evt. ekstra afhængigheder: VS Code, typecheck, ekstra QA ===
# pylint
# isort
# pyright
